requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain-openai==0.0.5
langchain-core==0.1.10
langgraph==0.0.20
pydantic==2.5.0
typing-extensions==4.8.0
requests==2.31.0
langsmith==0.0.69 